+++ src/mcp/command-tools.ts
@@ -10,7 +10,7 @@ import {
   listMcpCommandMetadata,
   type CommandName,
 } from '../commands/command-metadata.ts';
-import { mcpBody } from '../cli-schema/command-text.ts';
+import { mcpBody } from '../commands/command-text.ts';
 import { resolveCommandRecordsSessionAction } from '../core/command-descriptor/registry.ts';
 import { MCP_COMMAND_OUTPUT_SCHEMAS } from './mcp-output-schemas.ts';
 import { AppError } from '@agent-device/kernel/errors';
